Add support for helm diff

+++ index.ts
@@ -1,5 +1,6 @@
 // deno-lint-ignore-file
 import * as fs from "https://deno.land/std@0.86.0/fs/mod.ts"
+import { parseHelmArgs } from "./args/parse-helm-args.ts"
 import { parseArgs } from "./args/parse-helm-deno-args.ts"
 import { renderDenoChart } from "./lib/deno/index.ts"
 import {
@@ -10,8 +11,6 @@ import {
   ignoreNotFoundError,
 } from "./lib/helm/index.ts"
 
-const supportedHelmCommand = ["template", "install", "upgrade"]
-
 function helmDenoUsage() {
   const pluginUsage = `
 This is a wrapper for "helm [command]". It will use Deno for rendering charts
@@ -21,6 +20,7 @@ Supported helm [command] is:
   - template
   - install
   - upgrade
+  - diff (helm plugin)
 
 You must use the options of the supported commands in strict order:
   $ helm [command] [RELEASE] [CHART] [flags]
@@ -59,54 +59,29 @@ function withErrorMsg<T>(p: Promise<T>, msg: string): Promise<T> {
   return p.catch((err) => Promise.reject(`${msg}: ${err}`))
 }
 
-function getArgsWithoutPlugins(args: string[]): string[] {
-  // TODO: get current installed plugins
-  const plugins = ["conftest", "diff", "push", "secrets"]
-
-  for (let i = 0; i < args.length; i++) {
-    const arg = args[i]
-
-    if (arg.startsWith("-")) {
-      continue
-    }
-
-    if (plugins.some((plugin) => plugin === arg)) {
-      return getArgsWithoutPlugins([...args.slice(0, i - 1), ...args.slice(i)])
-    }
-
-    break
-  }
-
-  return args
-}
-
-function normalizeArgs(args: string[]): string[] {
-  return getArgsWithoutPlugins(args)
-}
-
 async function main() {
-  const { helmArgs: args, options } = parseArgs(normalizeArgs(Deno.args))
+  const { helmArgs: args, options } = parseArgs(Deno.args)
+  const startTime = Date.now()
   const debug: typeof console.error = (...args) => {
     if (options.logLevel === "debug") {
-      console.error(...args)
+      console.error(`[${Date.now() - startTime} ms]`, ...args)
     }
   }
   debug(`Running with options:\n${JSON.stringify(options, null, 2)}`)
 
-  const helmCommand = args[0]
-  const releaseName = args[1]
-  const chartPath = args[2]
-
-  if (!helmCommand || helmCommand === "-h" || helmCommand === "--help") {
+  if (!args[0] || args[0] === "-h" || args[0] === "--help") {
     helmDenoUsage()
     return
   }
 
-  if (!supportedHelmCommand.includes(helmCommand)) {
-    throw new Error(
-      `${helmCommand} command not supported, please check documentation.`
-    )
-  }
+  const {
+    command,
+    releaseName,
+    chartLocation,
+    options: helmRestArgs,
+  } = parseHelmArgs(args)
+
+  // TODO: validate command is allowed
 
   const workdir = await withErrorMsg(
     Deno.makeTempDir({ prefix: "chart-" }),
@@ -114,18 +89,18 @@ async function main() {
   )
 
   debug(`Temporary directory ${workdir} has been created`)
-  const isLocalChart = isChartExist(chartPath)
+  const isLocalChart = isChartExist(chartLocation)
 
   try {
     // Fetch chart into temporaty directory
     if (isLocalChart) {
-      debug(`Copying chart ${chartPath} to temporary directory`)
-      await copyChart(chartPath, workdir)
-      debug(`Successfuly copied chart ${chartPath} to temporary directory`)
+      debug(`Copying chart ${chartLocation} to temporary directory`)
+      await copyChart(chartLocation, workdir)
+      debug(`Successfuly copied chart ${chartLocation} to temporary directory`)
     } else {
-      debug(`Fetching chart ${chartPath} to temporary directory`)
-      await fetchChart(chartPath, workdir)
-      debug(`Successfuly fetched chart ${chartPath} to temporary directory`)
+      debug(`Fetching chart ${chartLocation} to temporary directory`)
+      await fetchChart(chartLocation, workdir)
+      debug(`Successfuly fetched chart ${chartLocation} to temporary directory`)
     }
 
     const chartContext = await getChartContext(releaseName, workdir, args)
@@ -134,23 +109,24 @@ async function main() {
     await renderDenoChart(chartContext, workdir)
     debug("Deno templates were successfuly rendered")
 
-    debug(
-      `Executing: ${helmCommand} ${releaseName} ${workdir} ${args
-        .slice(3)
-        .join(" ")}`
-    )
-    await helmExecute(helmCommand, releaseName, workdir, args.slice(3))
+    const helmExecuteArgs = [...command, releaseName, workdir, ...helmRestArgs]
+    debug(`Executing: ${helmExecuteArgs.join(" ")}`)
+    await helmExecute(helmExecuteArgs)
+
+    debug("Success")
   } catch (err) {
     if (err?.message) {
       // Replace paths in stacktrace with readable versions
       err.message = isLocalChart
-        ? err.message.replaceAll(workdir, chartPath)
+        ? err.message.replaceAll(workdir, chartLocation)
         : err.message.replaceAll(`file://${workdir}`, "<chart-root>")
     }
     throw err
   } finally {
-    // Remove temporary directory
-    await ignoreNotFoundError(Deno.remove(workdir, { recursive: true }))
+    if (!options.keepTmpChart) {
+      // Remove temporary directory
+      await ignoreNotFoundError(Deno.remove(workdir, { recursive: true }))
+    }
   }
 }
